refactor(lux-depth): split segmentation backend package
